Moved all main moodler functions into moodler/moodle.py
